Did part 1 of tech Challenge
created 2 columns that apear side by side.
Put in a header and footer.
Put in an unordered list in the second column.
Added 2 images
Added a link that goes to an outside resource. Labeled as "Blog on Edwin Pepping.
Used Semantic HTML with only a few div elements being used.